refactor: Add Fattoria class
This commit adds the Fattoria class, which represents a farm. It includes methods for adding and removing animals, as well as a method for printing the list of animals. This addition enhances the functionality of the codebase by providing a way to manage and display animals in a farm setting.